refactor(navbar): remove style folder, add style to component
